🚧chore: Make formal fork of `django-ltree`
**django-ltree** (`Github repository <https://github.com/mariocesar/django-ltree>`__) was
created in March 2020 by Mario-César. It went unmaintained from August 2021.

Kimsia Sim was motivated to fork it under the name **greendeploy-django-ltree** in September 2023
and make it support Django 4.2. The reason is that Kimsia needed to use it in a Django 4.2 project.

The Django package setup that this fork is emulating from is `django-cors-headers <https://github.com/adamchainz/django-cors-headers/tree/main>`__

Signed-off-by: KimSia, Sim <245021+simkimsia@users.noreply.github.com>